# static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InnoDD API",
    "version": "2.0.0",
    "description": "Complete InnoDD drug-discovery platform surface, generated from source (not the Postman collection).\n\n**Two surfaces:**\n- `/api/v1/*` — legacy internal pipeline modules (TxKG, Literature Mining, Drug Curation, Screening, Novelty Search) plus Auth/Users/Items scaffolding and job Orchestration. No X-Drp-Token needed.\n- `/v1/*` — the DRP frontend-facing contract (async job pattern: POST creates a job -> 202 {jobId}, then poll status/result). Needs `X-Drp-Token` (or `Authorization: Bearer`) from `/v1/auth/login`.\n\n**Known gaps / stubs (do not rely on for testing):**\n- `POST /api/v1/auth/login` and `/api/v1/auth/token` are unimplemented stubs (`pass`) in the legacy scaffold.\n- `/api/v1/users/*` imports a `UserRead` model that does not exist in `app/schemas/user.py` — this router likely fails to import and is NOT actually mounted.\n- `/api/v1/agentic/*` never mounts — `agent_routes.py` does not exist as a file; `_include()` swallows the ImportError.\n- Several module routers are double-mounted: Literature Mining, Drug Curation, Screening, and Novelty Search each exist BOTH under their versioned prefix (e.g. `/api/v1/literature-mining/...`) AND again with no prefix as a legacy alias (e.g. `/api/v1/search-keywords/`) — both paths are live simultaneously and hit the same handler.",
    "contact": { "name": "InnoDD API" }
  },
  "servers": [
...
  ],
  "components": {
    "securitySchemes": {
      "drpToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Drp-Token",
        "description": "Access token returned by POST /v1/auth/login or /v1/auth/refresh. The server also accepts a standard Authorization: Bearer header as a fallback."
      }
    },
    "schemas": {
      "JobAccepted": { "type": "object", "properties": { "jobId": { "type": "string" } } },
...
      "ItemCreate": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } } },
      "ItemUpdate": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } } },
      "Item": { "type": "object", "properties": { "id": { "type": "integer" }, "name": { "type": "string" }, "description": { "type": "string" } } },
...
      "ArticleResult": {
        "type": "object",
        "properties": {
          "protein_name": { "type": "string" },
          "title": { "type": "string" },
          "score": { "type": "number", "minimum": 0, "maximum": 100 },
          "pdf_file_path": { "type": "string", "nullable": true },
          "found_keywords": { "type": "array", "items": { "type": "string" } },
          "preview": { "type": "string" }
        }
      },
      "SearchResponse": {
        "type": "object",
        "properties": { "results": { "type": "array", "items": { "$ref": "#/components/schemas/ArticleResult" } } }
      },
      "LitMineXRequest": {
        "type": "object",
        "properties": {
          "target": { "type": "string", "nullable": true },
          "disease": { "type": "string", "nullable": true },
          "query": { "type": "string", "nullable": true },
          "source": { "type": "string", "enum": ["txkg_carryover", "user_direct"], "default": "user_direct" },
          "session_id": { "type": "string", "nullable": true },
          "max_results": { "type": "integer", "minimum": 1, "maximum": 500 },
          "top_n": { "type": "integer", "minimum": 1, "maximum": 50 },
          "exclude_reviews": { "type": "boolean" }
        },
        "description": "At least one of query/target/disease is required."
      },
      "LitMineXFollowUpRequest": {
        "type": "object",
        "required": ["question", "session_id"],
        "properties": {
          "question": { "type": "string" },
          "session_id": { "type": "string" },
          "pmids": { "type": "array", "items": { "type": "string" } },
          "deep": { "type": "boolean", "default": false }
        }
      },
      "LitMineXResponse": {
        "type": "object",
        "properties": {
          "module": { "type": "string", "example": "litminex" },
          "target": { "type": "string", "nullable": true },
          "disease": { "type": "string", "nullable": true },
          "query_answer": {
            "type": "object",
            "properties": { "text": { "type": "string" }, "cited_pmids": { "type": "array", "items": { "type": "string" } } }
          },
          "article_table": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "pmid": { "type": "string" }, "title": { "type": "string" }, "relevance_score": { "type": "number" },
                "pdf_link": { "type": "string" }, "journal": { "type": "string" }, "pub_date": { "type": "string" },
                "per_article_summary": { "type": "string" }
              }
            }
          },
          "session_state_update": {
            "type": "object",
            "properties": { "litminex_last_query": { "type": "string" }, "litminex_articles": { "type": "array", "items": { "type": "object" } } }
          },
          "ranked_articles": { "type": "array", "items": { "type": "object" } },
          "query_terms": { "type": "object" },
          "pubmed_query": { "type": "string" },
          "pipeline_backends": { "type": "object" }
        }
      },
...
      "ScreenCompound": { "type": "object", "properties": { "drug_name": { "type": "string" }, "sdf_file_path": { "type": "string" } } },
      "ScreenRequest": {
        "type": "object",
        "required": ["target"],
        "properties": {
          "target": { "type": "string" }, "compoundLibrary": { "type": "string", "nullable": true },
          "pdbFilePath": { "type": "string", "nullable": true }, "compounds": { "type": "array", "items": { "$ref": "#/components/schemas/ScreenCompound" } }
        }
      },
      "ScreeningHit": { "type": "object", "properties": { "mode": { "type": "integer" }, "compound": { "type": "string" }, "protein": { "type": "string" }, "affinityKcalPerMol": { "type": "number" }, "outputFile": { "type": "string" } } },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "name": { "type": "string" }, "disease": { "type": "string" },
          "module": { "type": "string" }, "status": { "type": "string" }, "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ProjectPage": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } },
          "totalCount": { "type": "integer" }, "page": { "type": "integer" }, "totalPages": { "type": "integer" }
        }
      },
      "CreateProjectRequest": {
        "type": "object",
        "required": ["name"],
        "properties": { "name": { "type": "string" }, "disease": { "type": "string", "default": "" }, "module": { "type": "string", "default": "" }, "status": { "type": "string", "default": "Active" } }
      },
      "ProjectItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "resultType": { "type": "string" }, "sessionId": { "type": "string", "nullable": true },
          "resultId": { "type": "string", "nullable": true }, "payload": { "type": "object" }, "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "SaveResultRequest": {
        "type": "object",
        "required": ["resultId", "resultType"],
        "properties": {
          "sessionId": { "type": "string", "nullable": true }, "resultId": { "type": "string" },
          "resultType": { "type": "string", "enum": ["targets", "subgraph", "metapath", "litminex_results", "article"] }
        }
      },
      "SaveResultResponse": { "type": "object", "properties": { "success": { "type": "boolean" }, "projectName": { "type": "string" }, "itemId": { "type": "string" } } },
...
    }
  },
...
  "paths": {
...
    "/api/v1/items/": {
      "post": {
        "tags": ["Items (legacy scaffold)"], "summary": "Create item",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ItemCreate" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } } }
      }
    },
    "/api/v1/items/{item_id}": {
      "get": { "tags": ["Items (legacy scaffold)"], "summary": "Read item", "parameters": [{ "name": "item_id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } } } },
      "put": {
        "tags": ["Items (legacy scaffold)"], "summary": "Update item",
        "parameters": [{ "name": "item_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ItemUpdate" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Item" } } } } }
      },
      "delete": { "tags": ["Items (legacy scaffold)"], "summary": "Delete item", "parameters": [{ "name": "item_id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "OK" } } }
    },
...
    "/api/v1/search-keywords/": {
      "get": {
        "tags": ["2. Literature Mining"], "summary": "Search keywords (legacy alias, no prefix)",
        "parameters": [
          { "name": "article_keywords", "in": "query", "required": true, "schema": { "type": "string" }, "example": "EGFR" },
          { "name": "search_keywords", "in": "query", "schema": { "type": "string", "default": "" } },
          { "name": "max_results", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1, "maximum": 100 } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SearchResponse" } } } } }
      }
    },
...
    "/api/v1/litminex/query": {
      "post": {
        "tags": ["2. Literature Mining"], "summary": "LitMineX query (legacy alias)",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LitMineXRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LitMineXResponse" } } } } }
      }
    },
...
    "/api/v1/litminex/followup": {
      "post": {
        "tags": ["2. Literature Mining"], "summary": "LitMineX follow-up QA (legacy alias)",
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LitMineXFollowUpRequest" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
...
    "/v1/agents/screensuite/screen": {
      "post": {
        "tags": ["DRP Agents - ScreenSuite"], "summary": "ScreenSuite - Screen (creates job + DrpPipeline row)", "security": [{ "drpToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ScreenRequest" } } } },
        "responses": { "202": { "description": "Job accepted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobAccepted" } } } } }
      }
    },
    "/v1/agents/screensuite/{jobId}/hits": {
      "get": {
        "tags": ["DRP Agents - ScreenSuite"], "summary": "ScreenSuite - Docking hits", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "jobId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ScreeningHit" } } } } } }
      }
    },
...
    "/v1/projects": {
      "get": {
        "tags": ["DRP Projects"], "summary": "List projects", "security": [{ "drpToken": [] }],
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "module", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string", "default": "Latest Activity" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectPage" } } } } }
      },
      "post": {
        "tags": ["DRP Projects"], "summary": "Create project", "security": [{ "drpToken": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateProjectRequest" } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } } }
      }
    },
    "/v1/projects/{projectId}": {
      "get": {
        "tags": ["DRP Projects"], "summary": "Get project", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "projectId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } } }
      }
    },
    "/v1/projects/{projectId}/items": {
      "get": {
        "tags": ["DRP Projects"], "summary": "List project items", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "projectId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProjectItem" } } } } } }
      }
    },
    "/v1/projects/{projectId}/results": {
      "post": {
        "tags": ["DRP Projects"], "summary": "Save a result into a project", "security": [{ "drpToken": [] }],
        "parameters": [{ "name": "projectId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveResultRequest" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SaveResultResponse" } } } } }
      }
    },
...
  }
}